Some random fixes for compiling under clang.  Will readd rt later.

+++ src/include/signal_tree/tree.h
@@ -83,7 +83,7 @@ namespace bcpp
     
 
         //=============================================================================
-        template <std::uint64_t N>
+        template <std::size_t N>
         class tree final
         {
         private:
